% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_asynchrony.R
\name{compute_asynchrony}
\alias{compute_asynchrony}
\title{Compute the asynchrony based on Tilted-CCA}
\usage{
compute_asynchrony(input_obj, anchor_modality = 1)
}
\arguments{
\item{input_obj}{\code{multiSVD} class, after applying \code{tiltedCCA_decomposition()}}

\item{anchor_modality}{numeric of \code{1} or \code{2}, denoting which modality the synchrony is computed on}
}
\value{
a matrix with \code{n} cells and 2 columns, which one column 
(named \code{asynchrony}) denotes the synchrony score (between 0 and 1, where a number closer to 1 means
the cell is synchronous between both modalities) and 
(named \code{asynchrony_rescaled}), which is simply a monotone transformation
of the first column for easier visualization
}
\description{
Compute the asynchrony based on Tilted-CCA
}
