% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_cell.R
\name{compute_enrichment_scores}
\alias{compute_enrichment_scores}
\title{Compute enrichment scores}
\usage{
compute_enrichment_scores(c_g, d_g_1, d_g_2, idx)
}
\arguments{
\item{c_g}{sparse matrix of class \code{dgCMatrix}, ideally from \code{combine_frnn}
representing the combined common embedding, where the non-zero entries represent distances}

\item{d_g_1}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the distinct embedding in one modality, where the non-zero entries represent distances}

\item{d_g_2}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the distinct embedding in the other modality, where the non-zero entries represent distances}

\item{idx}{initial indicies of cells-of-interest}
}
\value{
vector
}
\description{
Compute enrichment scores
}
