% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_laplacian.R
\name{compute_laplacian}
\alias{compute_laplacian}
\title{Compute the eigenbasis of a Laplacian based on a sparse matrix}
\usage{
compute_laplacian(
  g_mat,
  k_max = 100,
  normalize = T,
  rowname_vec,
  colname_vec,
  verbose = T
)
}
\arguments{
\item{g_mat}{sparse matrix of class \code{dgCMatrix}}

\item{k_max}{positive integer}

\item{normalize}{boolean}

\item{rowname_vec}{vector of rownames}

\item{colname_vec}{vector of colnames}

\item{verbose}{boolean}
}
\value{
matrix of dimension \code{nrow(g_mat)} by \code{k_max}
}
\description{
Compute the eigenbasis of a Laplacian based on a sparse matrix
}
