% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_laplacian.R
\name{compute_smooth_signal}
\alias{compute_smooth_signal}
\title{Compute smoothed vector based on eigenbases}
\usage{
compute_smooth_signal(vec, eigenbasis)
}
\arguments{
\item{vec}{vector}

\item{eigenbasis}{matrix with rows equal to \code{length(vec)}}
}
\value{
a list
}
\description{
Compute smoothed vector based on eigenbases
}
