% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snn.R
\name{compute_snns}
\alias{compute_snns}
\title{Include SNN graphs to multiSVD}
\usage{
compute_snns(
  input_obj,
  latent_k,
  num_neigh,
  bool_cosine = T,
  bool_intersect = T,
  min_deg = 1,
  tol = 1e-04,
  verbose = 0
)
}
\arguments{
\item{input_obj}{\code{multiSVD} class, after creation via \code{create_multiSVD()} or \code{form_metacells()}}

\item{latent_k}{number of latent dimensions for the graph Laplacian bases}

\item{num_neigh}{number of neighbors for each cell, when constructing the SNN graphs}

\item{bool_cosine}{boolean, for using cosine distance if \code{T} or Euclidean distance if \code{F}}

\item{bool_intersect}{boolean, on whether or not to symmetrize (via the AND function) the SNN}

\item{min_deg}{minimum degree of each cell in the SNN}

\item{tol}{small positive number}

\item{verbose}{non-negative integer}
}
\value{
updated \code{multiSVD} object
}
\description{
Computing and including the shared nearest-neighbor (SNN) graphs 
for each modality and the target common manifold
for an existing \code{multiSVD} object
}
