% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_pca.R
\name{consensus_pca}
\alias{consensus_pca}
\title{Computing Consensus PCA}
\usage{
consensus_pca(
  mat_1,
  mat_2,
  dims_1,
  dims_2,
  dims_consensus,
  apply_pca = T,
  center_1 = T,
  center_2 = T,
  center_consensus = T,
  normalize_row = F,
  normalize_singular_value = T,
  recenter_1 = F,
  recenter_2 = F,
  recenter_consensus = F,
  rescale_1 = F,
  rescale_2 = F,
  rescale_consensus = F,
  scale_1 = T,
  scale_2 = T,
  scale_consensus = T,
  scale_max_1 = NULL,
  scale_max_2 = NULL,
  scale_max_consensus = NULL,
  svd_1 = NULL,
  svd_2 = NULL,
  tol = 0.001,
  verbose = 0
)
}
\arguments{
\item{mat_1}{data matrix of \code{n} cells and \code{p1} features for Modality 1}

\item{mat_2}{data matrix of \code{n} cells and \code{p2} features for Modality 2}

\item{dims_1}{vector of latent dimensions for \code{mat_1} used for analysis}

\item{dims_2}{vector of latent dimensions for \code{mat_2} used for analysis}

\item{dims_consensus}{vector of latent dimensions for the Consensus PCA}

\item{apply_pca}{boolean, where PCA is combined set of latent dimensions (from both modalities) if \code{T}
or not if \code{F}}

\item{center_1}{boolean, to center each the feature in Modality 1 prior to computing latent dimensions}

\item{center_2}{boolean, to center each the feature in Modality 2 prior to computing latent dimensions}

\item{center_consensus}{boolean, to center the combined set of latent dimensions (from both modalities)}

\item{normalize_row}{boolean, to normalize each cell's latent vector after dimension-reduction for both modalities}

\item{normalize_singular_value}{boolean, to normalize each modality by its largest singular value}

\item{recenter_1}{boolean, to center each latent dimension in Modality 1 after computing latent dimensions}

\item{recenter_2}{boolean, to center each latent dimension in Modality 2 after computing latent dimensions}

\item{recenter_consensus}{boolean, to center the latent dimension after Consensus PCA}

\item{rescale_1}{boolean, to rescale each latent dimension in Modality 1 after computing latent dimensions}

\item{rescale_2}{boolean, to rescale each latent dimension in Modality 2 after computing latent dimensions}

\item{rescale_consensus}{boolean, to rescale the latent dimension after Consensus PCA}

\item{scale_1}{boolean, to rescale each the feature in Modality 1 prior to computing latent dimensions}

\item{scale_2}{boolean, to rescale each the feature in Modality 2 prior to computing latent dimensions}

\item{scale_consensus}{boolean, to rescale the combined set of latent dimensions (from both modalities)}

\item{scale_max_1}{numeric or \code{NULL}, to threshold Modality 1 in magnitude prior to computing latent dimensions}

\item{scale_max_2}{numeric or \code{NULL}, to threshold Modality 2 in magnitude prior to computing latent dimensions}

\item{scale_max_consensus}{numeric or \code{NULL}, to threshold the combined set of latent dimensions in magnitude prior to computing latent dimensions}

\item{svd_1}{list of \code{u}, \code{d}, \code{v} for the SVD of Modality 1 if it's already computed}

\item{svd_2}{list of \code{u}, \code{d}, \code{v} for the SVD of Modality 2 if it's already computed}

\item{tol}{small positive number}

\item{verbose}{non-negative integer}
}
\value{
\code{consensusPCA} object
}
\description{
Computing Consensus PCA
}
