% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_celltype_subsample}
\alias{construct_celltype_subsample}
\title{Construct cell-type subsamples}
\usage{
construct_celltype_subsample(membership_vec, min_subsample_cell)
}
\arguments{
\item{membership_vec}{factor vector}

\item{min_subsample_cell}{positive integer}
}
\value{
indices
}
\description{
Sample at most \code{min_subsample_cell} cells from each
unique level of \code{membership_vec}
}
