% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clisi_information.R
\name{construct_frnn}
\alias{construct_frnn}
\title{Construct fixed-radius NN graph}
\usage{
construct_frnn(
  common_score,
  distinct_score,
  svd_e,
  cell_subidx,
  nn,
  frnn_approx = 0,
  radius_quantile = 0.9,
  bool_matrix = F,
  verbose = T
)
}
\arguments{
\item{common_score}{output from \code{dcca_decomposition}}

\item{distinct_score}{output from \code{dcca_decomposition}}

\item{svd_e}{list containing the SVD of the full matrix}

\item{cell_subidx}{vector of integers between 1 and \code{nrow(common_score)}}

\item{nn}{integer of number of nearest neighbors to determine the appropriate radius
for the frNN graph}

\item{frnn_approx}{small non-negative number}

\item{radius_quantile}{value between 0 and 1}

\item{bool_matrix}{boolean. If \code{TRUE}, output the graphs as a sparse matrix.
If \code{FALSE}, output the graphs as a list where each element of the list
corresponds with the element's neighbors}

\item{verbose}{boolean}
}
\value{
list, depends on \code{bool_matrix}
}
\description{
Construct fixed-radius NN graph
}
