% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.R
\name{create_multiSVD}
\alias{create_multiSVD}
\title{Create the object to initialize Tilted-CCA}
\usage{
create_multiSVD(
  mat_1,
  mat_2,
  dims_1,
  dims_2,
  center_1 = T,
  center_2 = T,
  normalize_row = T,
  normalize_singular_value = T,
  recenter_1 = F,
  recenter_2 = F,
  rescale_1 = F,
  rescale_2 = F,
  scale_1 = T,
  scale_2 = T,
  scale_max_1 = NULL,
  scale_max_2 = NULL,
  verbose = 0
)
}
\arguments{
\item{mat_1}{data matrix of \code{n} cells and \code{p1} features for Modality 1}

\item{mat_2}{data matrix of \code{n} cells and \code{p2} features for Modality 2}

\item{dims_1}{vector of latent dimensions for \code{mat_1} used for analysis}

\item{dims_2}{vector of latent dimensions for \code{mat_2} used for analysis}

\item{center_1}{boolean, to center each the feature in Modality 1 prior to computing latent dimensions}

\item{center_2}{boolean, to center each the feature in Modality 2 prior to computing latent dimensions}

\item{normalize_row}{boolean, to normalize each cell's latent vector after dimension-reduction for both modalities}

\item{normalize_singular_value}{boolean, to normalize each modality by its largest singular value}

\item{recenter_1}{boolean, to center each latent dimension in Modality 1 after computing latent dimensions}

\item{recenter_2}{boolean, to center each latent dimension in Modality 2 after computing latent dimensions}

\item{rescale_1}{boolean, to rescale each latent dimension in Modality 1 after computing latent dimensions}

\item{rescale_2}{boolean, to rescale each latent dimension in Modality 2 after computing latent dimensions}

\item{scale_1}{boolean, to rescale each the feature in Modality 1 prior to computing latent dimensions}

\item{scale_2}{boolean, to rescale each the feature in Modality 2 prior to computing latent dimensions}

\item{scale_max_1}{numeric or \code{NULL}, to threshold Modality 1 in magnitude prior to computing latent dimensions}

\item{scale_max_2}{numeric or \code{NULL}, to threshold Modality 2 in magnitude prior to computing latent dimensions}

\item{verbose}{non-negative integer}
}
\value{
\code{multiSVD} object
}
\description{
Initialize the \code{multiSVD} object by computing each modality's SVD.
}
