% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helpers.R
\name{create_SeuratDim}
\alias{create_SeuratDim}
\title{Create Seurat dimension reduction objects of Tilted-CCA}
\usage{
create_SeuratDim(
  input_obj,
  what,
  aligned_umap_assay = NULL,
  scale_max_1 = NULL,
  scale_max_2 = NULL,
  seurat_obj = NULL,
  seurat_assay = "RNA",
  suppress_warnings = TRUE,
  verbose = 0,
  ...
)
}
\arguments{
\item{input_obj}{a \code{multiSVD} object that is the output of \code{tiltedCCA_decomposition}}

\item{what}{a character, either \code{"common"}, \code{"distinct_1"}, or \code{"distinct_2"} for 
which embedding to construct the visualization for}

\item{aligned_umap_assay}{either \code{NULL} or a UMAP assay in \code{seurat_assay} 
(in which case the resulting UMAP will be rotated to best mimic the relative orientation
of cells in \code{seurat_obj[[aligned_umap_assay]]})}

\item{scale_max_1}{numeric or \code{NULL}, to threshold Modality 1 in magnitude prior to computing latent dimensions}

\item{scale_max_2}{numeric or \code{NULL}, to threshold Modality 2 in magnitude prior to computing latent dimensions}

\item{seurat_obj}{the \code{Seurat} object that was used to compute \code{input_obj}, the \code{multiSVD_obj}}

\item{seurat_assay}{the assay in \code{seurat_obj} to assign the resulting embedding to}

\item{suppress_warnings}{boolean to suppress the warning when running \code{Seurat::RunUMAP}}

\item{verbose}{non-negative integer}

\item{...}{extra arguments to \code{Seurat::RunUMAP}}
}
\value{
a \code{DimReduc} \code{SeuratObject}
}
\description{
Create Seurat dimension reduction objects of Tilted-CCA
}
