% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{dcca}
\alias{dcca}
\title{D-CCA}
\usage{
dcca(mat_1, mat_2, rank_1, rank_2, rank_12, enforce_rank = T, verbose = T)
}
\arguments{
\item{mat_1}{data matrix 1}

\item{mat_2}{data matrix 2}

\item{rank_1}{desired rank of data matrix 1}

\item{rank_2}{desired rank of data matrix 1}

\item{rank_12}{desired rank of cross-covariance matrix}

\item{enforce_rank}{boolean}

\item{verbose}{boolean}
}
\value{
list
}
\description{
D-CCA
}
