% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{dcca_factor}
\alias{dcca_factor}
\title{D-CCA}
\usage{
dcca_factor(mat_1, mat_2, rank_1, rank_2, enforce_rank = T, verbose = T)
}
\arguments{
\item{mat_1}{data matrix 1}

\item{mat_2}{data matrix 2}

\item{rank_1}{desired rank of data matrix 1}

\item{rank_2}{desired rank of data matrix 1}

\item{enforce_rank}{boolean}

\item{verbose}{boolean}
}
\value{
list
}
\description{
D-CCA
}
