% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_information_weight.R
\name{dcca_information_weight}
\alias{dcca_information_weight}
\title{Weight based on D-CCA}
\usage{
dcca_information_weight(
  mat,
  common_mat,
  tol = 0.05,
  iter_max = 100,
  verbose = T
)
}
\arguments{
\item{mat}{Desired observed matrix}

\item{common_mat}{Desired common matrix}

\item{tol}{numeric}

\item{iter_max}{numeric}

\item{verbose}{boolean}
}
\value{
three vectors as a list
}
\description{
Weight based on D-CCA
}
