% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_explained_variance.R
\name{dcca_variance_decomposition}
\alias{dcca_variance_decomposition}
\title{D-CCA Variance decomposition}
\usage{
dcca_variance_decomposition(dcca_res, rank_c, verbose = T)
}
\arguments{
\item{dcca_res}{Output of \code{dcca_factor}}

\item{rank_c}{desired rank of cross-correlation matrix between \code{mat_1} and \code{mat_2} when running \code{dcca_factor}}

\item{verbose}{boolean}
}
\value{
object of class \code{dcca_decomp2}
}
\description{
D-CCA Variance decomposition
}
