% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_cell.R
\name{detect_rare_cell}
\alias{detect_rare_cell}
\title{Detect rare cells}
\usage{
detect_rare_cell(
  c_g,
  d_g_1,
  d_g_2,
  idx,
  common_enrich,
  distinct_enrich_1,
  distinct_enrich_2,
  custom_threshold = rep(NA, 3),
  deg_threshold = 0,
  max_size = NA,
  tol = 0.02,
  max_tries = 10,
  verbose = F
)
}
\arguments{
\item{c_g}{sparse matrix of class \code{dgCMatrix}, ideally from \code{combine_frnn}
representing the combined common embedding, where the non-zero entries represent distances}

\item{d_g_1}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the distinct embedding in one modality, where the non-zero entries represent distances}

\item{d_g_2}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the distinct embedding in the other modality, where the non-zero entries represent distances}

\item{idx}{initial indicies of cells-of-interest}

\item{common_enrich}{boolean, where \code{TRUE} means you want to find a set of cells
with a high common enrichment}

\item{distinct_enrich_1}{boolean, where \code{TRUE} means you want to find a set of cells
with a high distinct enrichment in one modality}

\item{distinct_enrich_2}{boolean, where \code{TRUE} means you want to find a set of cells
with a high distinct enrichment in the other modality}

\item{custom_threshold}{vector of length 3, where each entry is either \code{NA}
or a value between \code{0} and \code{1} (inclusive)}

\item{deg_threshold}{for a cell to be included into the set, it must be
neighbors with at least \code{deg_threshold} of the cells 
already in the set}

\item{max_size}{maximum size of the desired set}

\item{tol}{small positive integer}

\item{max_tries}{positive integer}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Detect rare cells
}
