% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Compute the differential expression across each pair of cell types}
\usage{
differential_expression(
  seurat_obj,
  assay,
  idents,
  slot = "data",
  test_use = "wilcox",
  verbose = T
)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{assay}{name of assay in the \code{seurat_obj}}

\item{idents}{variables name inside \code{seuart_obj} that contains the cell-type information (or clustering information) for each cell}

\item{slot}{slot of \code{seurat_obj[[assay]]} that informs with data matrix will be used in the DE test}

\item{test_use}{either \code{"MAST"} or \code{"wilcox"} dictates which DE test will be used}

\item{verbose}{non-negative integer}
}
\value{
a list with elements \code{"combn_mat"}, \code{"de_list"} and
\code{"level_vec"}
}
\description{
Compute the differential expression across each pair of cell types
}
