% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmca.R
\name{dmca_decomposition}
\alias{dmca_decomposition}
\title{D-MCA Decomposition}
\usage{
dmca_decomposition(dmca_res, verbose = T)
}
\arguments{
\item{dmca_res}{output from \code{dmca_factor}}

\item{verbose}{boolean}
}
\value{
list of class \code{dmca_decomp}
}
\description{
D-MCA Decomposition
}
