% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmca.R
\name{dmca_factor}
\alias{dmca_factor}
\title{D-MCA Factorization}
\usage{
dmca_factor(mat_1, mat_2, rank_1, rank_2, apply_shrinkage = T, verbose = T)
}
\arguments{
\item{mat_1}{data matrix 1}

\item{mat_2}{data matrix 2}

\item{rank_1}{desired rank of data matrix 1}

\item{rank_2}{desired rank of data matrix 2}

\item{apply_shrinkage}{boolean}

\item{verbose}{boolean}
}
\value{
list of class \code{dmca}
}
\description{
D-MCA Factorization
}
