% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{.compute_cca_aggregate_matrix}
\alias{.compute_cca_aggregate_matrix}
\title{Helper function with the CCA function}
\usage{
.compute_cca_aggregate_matrix(svd_1, svd_2, augment = T)
}
\arguments{
\item{svd_1}{SVD of the denoised variant of \code{mat_1} from \code{dcca_factor}}

\item{svd_2}{SVD of the denoised variant of \code{mat_2} from \code{dcca_factor}}

\item{augment}{boolean. If \code{TRUE}, augment the matrix with either rows or columns
with 0's so the dimension of the output matrix matches those in \code{svd_1} and \code{svd_2}}
}
\value{
matrix
}
\description{
Called by the \code{.cca} function. Recall when computing CCA,
the main matrix we need compute is, roughly speaking, 
half-inverse of the first covariance times the cross-covariance matrix
times the half-inverse of the second covariance. If we had the SVD
of the two original matrices, this is actually equivalent to the product
of the left singular vectors.
}
