% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{.compute_distinct_score}
\alias{.compute_distinct_score}
\title{Compute the distinct scores}
\usage{
.compute_distinct_score(score_1, score_2, common_score)
}
\arguments{
\item{score_1}{matrix}

\item{score_2}{matrix}

\item{common_score}{matrix}
}
\value{
list of two matrices
}
\description{
Given \code{score_1} and \code{score_2}, and having already
computed \code{common_score}, compute the distinct scores.
This is more-or-less a simple subtraction, but we need to handle situations
where we might need to "fill-in extra dimensions"
}
