% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variance.R
\name{.compute_pairwise_purity}
\alias{.compute_pairwise_purity}
\title{Compute pairwise purity}
\usage{
.compute_pairwise_purity(g, idx1, idx2, membership_vec)
}
\arguments{
\item{g}{\code{igraph} object}

\item{idx1}{index between 1 and \code{igraph::vcount(g)}}

\item{idx2}{index between 1 and \code{igraph::vcount(g)} not equal to \code{idx1}}

\item{membership_vec}{integer vector}
}
\value{
numeric
}
\description{
For two indices \code{idx1} and \code{idx2} (two nodes in the \code{igraph}
object \code{g}, so both \code{idx1} and \code{idx2} need to be less than
\code{igraph::vcount(g)}), compute the shortest path from node \code{idx1}
to node \code{idx2} and determine what percentage of nodes in that
path also share the same cluster label as node \code{idx1}
and node \code{idx2}.
}
