% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{.compute_prob_mat}
\alias{.compute_prob_mat}
\title{Compute probability matrix}
\usage{
.compute_prob_mat(B_mat, membership_vec)
}
\arguments{
\item{B_mat}{symmetric connectivity matrix}

\item{membership_vec}{vector containing values \code{1} through \code{ncol(B_mat)}}
}
\value{
symmetric matrix of dimension \code{length(membership_vec)}
}
\description{
Compute probability matrix
}
