% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{.compute_unnormalized_scores}
\alias{.compute_unnormalized_scores}
\title{Using the CCA solution, compute the score matrices.}
\usage{
.compute_unnormalized_scores(svd_1, svd_2, cca_res)
}
\arguments{
\item{svd_1}{SVD of the denoised variant of \code{mat_1} from \code{dcca_factor}}

\item{svd_2}{SVD of the denoised variant of \code{mat_2} from \code{dcca_factor}}

\item{cca_res}{returned object from \code{.cca}}
}
\value{
list of two matrices
}
\description{
This is called by \code{.tiltedCCA_common_score}. It's called unnormalized
scores since if there are \code{n} rows (i.e., \code{nrow(svd_1$u)}),
then the return matrices will be orthogonal matrices where the
matrix multiplied by itself is a diagonal matrix with all values \code{n}.
}
