% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softImpute.R
\name{.construct_missing_values}
\alias{.construct_missing_values}
\title{Construct indices for missing values}
\usage{
.construct_missing_values(n, p, num_val = ceiling(min(c(4, n/10, p/10))))
}
\arguments{
\item{n}{number of rows of intended matrix}

\item{p}{number of columns of intended matrix}

\item{num_val}{number of values to make missing in each row/column}
}
\value{
vector of indices between 1 and \code{n*p}
}
\description{
Construct indices for missing values
}
