% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{.dcca_common_score}
\alias{.dcca_common_score}
\title{Main workhorse of dcca_factor}
\usage{
.dcca_common_score(
  svd_1,
  svd_2,
  cca_res,
  num_neigh = max(round(nrow(svd_1$u)/20), 40),
  check_alignment = T,
  verbose = T,
  msg = ""
)
}
\arguments{
\item{svd_1}{SVD of the denoised variant of \code{mat_1} from \code{dcca_factor}}

\item{svd_2}{SVD of the denoised variant of \code{mat_2} from \code{dcca_factor}}

\item{cca_res}{returned object from \code{.cca}}

\item{num_neigh}{number of neighbors to consider to computed the common percentage}

\item{check_alignment}{boolean. If \code{TRUE}, recompute \code{score_1} and \code{score_2}
after using \code{.compute_unnormalized_scores}. This might be needed if the \code{.cca} solution
was not computed from exactly \code{svd_1} and \code{svd_2}}

\item{verbose}{boolean}

\item{msg}{character}
}
\value{
list
}
\description{
Given the two matrices (given by \code{svd_1} and \code{svd_2}) and the
CCA solution in \code{cca_res}, compute the common scores.
This calls the functions
\code{.common_decomposition} and \code{.compute_distinct_score}.
}
