% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{.extract_svd_embedding}
\alias{.extract_svd_embedding}
\title{Extract SVD embedding}
\usage{
.extract_svd_embedding(obj)
}
\arguments{
\item{obj}{output from either \code{generate_data} or \code{dcca_decomposition}}
}
\value{
list
}
\description{
Extract SVD embedding
}
