% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{.extract_umap_embedding}
\alias{.extract_umap_embedding}
\title{Extract UMAP embedding}
\usage{
.extract_umap_embedding(
  prep_list,
  common_1,
  common_2,
  distinct_1,
  distinct_2,
  add_noise,
  only_embedding,
  reduction_key = "UMAP"
)
}
\arguments{
\item{prep_list}{return object from \code{.prepare_umap_embedding}}

\item{common_1}{boolean}

\item{common_2}{boolean}

\item{distinct_1}{boolean}

\item{distinct_2}{boolean}

\item{add_noise}{boolean, intended (if \code{TRUE}) to put the common and 
distinct scores "on the same scale" by adding appropriately-scaled Gaussian noise}

\item{only_embedding}{boolean}

\item{reduction_key}{string for \code{Seurat::RunUMAP}}
}
\value{
2-column matrix or \code{Seurat} object
}
\description{
Extract UMAP embedding
}
