% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{.generate_adjaceny_mat}
\alias{.generate_adjaceny_mat}
\title{Simulate adjacency matrix}
\usage{
.generate_adjaceny_mat(prob_mat)
}
\arguments{
\item{prob_mat}{probability matrix}
}
\value{
adjacency matrix
}
\description{
The matrix is not symmetric.
}
