% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{.orthogonal_vec2vec}
\alias{.orthogonal_vec2vec}
\title{Projection of vector onto another vector}
\usage{
.orthogonal_vec2vec(vec1, vec2, tol = 1e-06)
}
\arguments{
\item{vec1}{vector}

\item{vec2}{vector}

\item{tol}{small positive number}
}
\value{
vector
}
\description{
Returns the component of \code{vec1} that is orthogonal to \code{vec2}
}
