% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.orthogonal_vector}
\alias{.orthogonal_vector}
\title{Projection of vector onto rows of a matrix}
\usage{
.orthogonal_vector(vec, mat)
}
\arguments{
\item{vec}{vector}

\item{mat}{matrix}
}
\value{
vector
}
\description{
Returns the component of \code{vec} that is orthogonal to all the rows
of \code{mat}
}
