% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{.prepare_umap_embedding}
\alias{.prepare_umap_embedding}
\title{Prepare the appropriate inputs for \code{.extract_umap_embedding}}
\usage{
.prepare_umap_embedding(obj)
}
\arguments{
\item{obj}{output from either \code{generate_data} or \code{dcca_decomposition}}
}
\value{
a list
}
\description{
Prepare the appropriate inputs for \code{.extract_umap_embedding}
}
