% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{.project_vec2vec}
\alias{.project_vec2vec}
\title{Projection of vector onto another vector}
\usage{
.project_vec2vec(vec1, vec2, orthogonal = T, tol = 1e-06)
}
\arguments{
\item{vec1}{vector}

\item{vec2}{vector}

\item{orthogonal}{boolean}

\item{tol}{small positive number}
}
\value{
vector
}
\description{
Returns the component of \code{vec1} that is orthogonal to \code{vec2}
if \code{orthogonal} is \code{TRUE}, and the
component of \code{vec1} that is parallel to \code{vec2} if 
\code{orthogonal} is \code{FALSE}.
}
