% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epca.R
\name{epca}
\alias{epca}
\title{Exponential PCA}
\usage{
epca(mat, K, mean_var_func = function(x) {
    x
}, verbose = T)
}
\arguments{
\item{mat}{data matrix}

\item{K}{desired rank}

\item{mean_var_func}{function relating mean to variance (depends on the distribution family being assumed)}

\item{verbose}{boolean}
}
\value{
set of \code{K} eigenvectors
}
\description{
Exponential PCA
}
