% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variance.R
\name{explained_variance}
\alias{explained_variance}
\title{Weights for cells and variables}
\usage{
explained_variance(dcca_decomp, membership_vec, num_pairs = 200)
}
\arguments{
\item{dcca_decomp}{Output of \code{dcca_variance_decomposition}}

\item{membership_vec}{integer vector}
}
\value{
4 vectors as a list
}
\description{
Weights for cells and variables
}
