% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{extract_umap_embedding}
\alias{extract_umap_embedding}
\title{Extract UMAP embedding}
\usage{
extract_umap_embedding(
  svd_list,
  common_1 = T,
  common_2 = T,
  distinct_1 = T,
  distinct_2 = T,
  noise_val = 0.05,
  only_embedding = T,
  reduction_key = "UMAP"
)
}
\arguments{
\item{svd_list}{return object from \code{extract_svd_embedding}}

\item{common_1}{boolean}

\item{common_2}{boolean}

\item{distinct_1}{boolean}

\item{distinct_2}{boolean}

\item{noise_val}{non-negative numeric (set to 0 to essentially ignore)}

\item{only_embedding}{boolean}

\item{reduction_key}{string for \code{Seurat::RunUMAP}}
}
\value{
2-column matrix or \code{Seurat} object
}
\description{
Extract UMAP embedding
}
