% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{fine_tuning}
\alias{fine_tuning}
\title{Fine tune the common tilts, one for each latent dimension}
\usage{
fine_tuning(
  input_obj,
  max_iter = 5,
  fix_tilt_perc = NA,
  temp_path = NULL,
  tol = 1e-05,
  verbose = 0
)
}
\arguments{
\item{input_obj}{\code{multiSVD} class, after creation via \code{tiltedCCA()}}

\item{max_iter}{maximum number of epochs (i.e., cycling through all latent dimensions)}

\item{fix_tilt_perc}{scalar between 0 and 1, or \code{NA}, for setting all the tilts in each latent dimension
to the scalar if not \code{NA}}

\item{temp_path}{filepath for saving temporary progress to}

\item{tol}{small positive number}

\item{verbose}{non-negative integer}
}
\value{
updated \code{multiSVD} object
}
\description{
Tune each latent dimension (representing a pair of canonical score vectors)
for an appropriate tilt of the common component
}
