% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{form_metacells}
\alias{form_metacells}
\title{Include meta-cell information to multiSVD}
\usage{
form_metacells(
  input_obj,
  large_clustering_1,
  large_clustering_2,
  num_metacells = NULL,
  min_size = 5,
  verbose = 0
)
}
\arguments{
\item{input_obj}{\code{multiSVD} class, after creation via \code{create_multiSVD()}}

\item{large_clustering_1}{factor among the \code{n} cells or \code{NULL}, representing the hard clustering structure of Modality 1}

\item{large_clustering_2}{factor among the \code{n} cells or \code{NULL}, representing the hard clustering structure of Modality 2}

\item{num_metacells}{number of desired meta-cells}

\item{min_size}{smallest number of cells for a meta-cell}

\item{verbose}{non-negative integer}
}
\value{
updated \code{multiSVD} object
}
\description{
Include the hard clustering information and compute the meta-cells 
for an existing  \code{multiSVD} object.
}
