% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  score_1,
  score_2,
  coef_mat_1,
  coef_mat_2,
  noise_func = function(mat) {     matrix(stats::rnorm(prod(dim(mat)), mean = mat),
    nrow(mat), ncol(mat)) }
)
}
\arguments{
\item{score_1}{\code{n} by \code{rank_1} orthogonal matrix}

\item{score_2}{\code{n} by \code{rank_2} orthogonal matrix}

\item{coef_mat_1}{\code{rank_1} by \code{p_1} matrix}

\item{coef_mat_2}{\code{rank_2} by \code{p_2} matrix}

\item{noise_func}{function that takes in a matrix and outputs a matrix of the same dimension}
}
\value{
list of two matrices, one of dimension \code{n} by \code{p_1} and another of dimension \code{n} by \code{p_2}
}
\description{
Generate data
}
