% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_random_orthogonal}
\alias{generate_random_orthogonal}
\title{Generate orthogonal matrices via Gaussian noise}
\usage{
generate_random_orthogonal(n, K, centered = F)
}
\arguments{
\item{n}{integer}

\item{K}{integer}

\item{centered}{boolean, where if \code{TRUE}, one less dimension
is provided but the singular vectors are centered}
}
\value{
matrix
}
\description{
Generate orthogonal matrices via Gaussian noise
}
