% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_sbm_orthogonal}
\alias{generate_sbm_orthogonal}
\title{Generate orthogonal matrices correspond to an SBM}
\usage{
generate_sbm_orthogonal(B_mat, membership_vec, centered = T)
}
\arguments{
\item{B_mat}{symmetric matrix}

\item{membership_vec}{vector of positive integers}

\item{centered}{boolean, where if \code{TRUE}, one less dimension
is provided but the singular vectors are centered}
}
\value{
matrix
}
\description{
Generate orthogonal matrices correspond to an SBM
}
