% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroPCA.R
\name{heteroPCA}
\alias{heteroPCA}
\title{HeteroPCA}
\usage{
heteroPCA(mat, K, max_iter = 25, tol = 1e-04, verbose = T)
}
\arguments{
\item{mat}{data matrix}

\item{K}{desired rank}

\item{max_iter}{numeric}

\item{tol}{numeric}

\item{verbose}{boolean}
}
\value{
covariance matrix
}
\description{
HeteroPCA
}
