% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_plot.R
\name{information_plot}
\alias{information_plot}
\title{Constructing information plot}
\usage{
information_plot(
  weight_mat,
  main = "",
  reorder_types = T,
  plot_individual = T,
  col_vec = (scales::hue_pal())(length(unique(weight_mat[, 3]))),
  y_offset = 0.1,
  cex_text = 1
)
}
\arguments{
\item{weight_mat}{data frame with 3 columns, where the first two columns
denote the row-specific weight for distinct information (values between 0
and 1) and the third column is the label}

\item{main}{character}

\item{reorder_types}{boolean}

\item{plot_individual}{boolean}

\item{col_vec}{vector of colors}

\item{y_offset}{numeric}

\item{cex_text}{numeric}
}
\value{
none
}
\description{
\code{weight_mat} has one row for each unit (ex: cell)
}
