% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Plot alignment of variables}
\usage{
plot_alignment(
  rsquare_vec,
  logpval_vec,
  bool_hide_points = F,
  bool_mark_ymedian = F,
  bool_polygon_mean = T,
  bool_truncate_xaxis = T,
  bool_white_bg = T,
  bty = "n",
  cex_axis = 1,
  cex_gene_highlight_inner = 3,
  cex_gene_highlight_outer = 4,
  cex_lab = 1,
  cex_points = 2,
  col_points = grDevices::rgb(0.5, 0.5, 0.5, 0.1),
  col_gene_highlight = 2,
  col_gene_highlight_border = "white",
  col_grid = "gray",
  density = 20,
  gene_names = NULL,
  lty_grid_major = 2,
  lty_grid_minor = 3,
  lty_polygon = 2,
  lwd_axis = 1,
  lwd_axis_ticks = 1,
  lwd_grid_major = 1,
  lwd_grid_minor = 0.5,
  lwd_polygon = 1,
  lwd_polygon_bold = 2,
  mark_median_xthres = 1,
  pch = 16,
  xaxt_num_ticks = 10,
  xaxt_grid_spacing = 2,
  xlab = "Separability (Median -Log10(p-value))",
  xlim = NULL,
  yaxt_by = 0.1,
  yaxt_num_ticks = 2,
  ylab = "Alignment with common space (R^2)",
  ylim = NULL,
  verbose = T,
  ...
)
}
\arguments{
\item{rsquare_vec}{output of \code{tiltedCCA::postprocess_modality_alignment}}

\item{logpval_vec}{output of \code{tiltedCCA::postprocess_depvalue}}

\item{bool_hide_points}{boolean (graphical parameter)}

\item{bool_mark_ymedian}{boolean (graphical parameter)}

\item{bool_polygon_mean}{boolean (graphical parameter)}

\item{bool_truncate_xaxis}{boolean (graphical parameter)}

\item{bool_white_bg}{boolean (graphical parameter)}

\item{bty}{character (graphical parameter)}

\item{cex_axis}{positive number (graphical parameter)}

\item{cex_gene_highlight_inner}{positive number (graphical parameter)}

\item{cex_gene_highlight_outer}{positive number (graphical parameter)}

\item{cex_lab}{positive number (graphical parameter)}

\item{cex_points}{positive number (graphical parameter)}

\item{col_points}{character for a color (graphical parameter)}

\item{col_gene_highlight}{character for a color (graphical parameter)}

\item{col_gene_highlight_border}{character for a color (graphical parameter)}

\item{col_grid}{character for a color (graphical parameter)}

\item{density}{positive number, \code{NA}, or \code{NULL} (graphical parameter)}

\item{gene_names}{vector of charaters for which genes (in \code{names(rsquare_vec)}) to highlight}

\item{lty_grid_major}{positive number (graphical parameter)}

\item{lty_grid_minor}{positive number (graphical parameter)}

\item{lty_polygon}{positive number (graphical parameter)}

\item{lwd_axis}{positive number (graphical parameter)}

\item{lwd_axis_ticks}{positive number (graphical parameter)}

\item{lwd_grid_major}{positive number (graphical parameter)}

\item{lwd_grid_minor}{positive number (graphical parameter)}

\item{lwd_polygon}{positive number (graphical parameter)}

\item{lwd_polygon_bold}{positive number (graphical parameter)}

\item{mark_median_xthres}{boolean (graphical parameter)}

\item{pch}{positive integer (graphical parameter)}

\item{xaxt_num_ticks}{positive integer (graphical parameter)}

\item{xaxt_grid_spacing}{positive integer (graphical parameter)}

\item{xlab}{character (graphical parameter)}

\item{xlim}{vector of 2 numerics, or \code{NULL} (graphical parameter)}

\item{yaxt_by}{positive number (graphical parameter)}

\item{yaxt_num_ticks}{positive integer (graphical parameter)}

\item{ylab}{character (graphical parameter)}

\item{ylim}{vector of 2 numerics, or \code{NULL} (graphical parameter)}

\item{verbose}{non-negative integer}

\item{...}{}
}
\value{
makes a plot but does not return anything
}
\description{
Plot alignment of variables
}
