% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_enrichment.R
\name{plot_cell_enrichment}
\alias{plot_cell_enrichment}
\title{Plot cell enrichment}
\usage{
plot_cell_enrichment(
  cell_enrichment_res,
  col_palette,
  cex_axis = 1,
  cex_lab = 1,
  lwd_diag = 3,
  lwd_grid = 2,
  mar = c(4, 5, 0.5, 0.5),
  xlab_1 = "Modality 1 Distinct",
  xlab_2 = "Modality 2 Distinct",
  ylab = "Common"
)
}
\arguments{
\item{cell_enrichment_res}{the result of \code{tiltedCCA::postprocess_cell_enrichment}}

\item{col_palette}{vector of colors (as strings), equal to the number of celltypes in \code{cell_enrichment_res}}

\item{cex_axis}{positive number (graphical parameter)}

\item{cex_lab}{positive number (graphical parameter)}

\item{lwd_diag}{positive number (graphical parameter)}

\item{lwd_grid}{positive number (graphical parameter)}

\item{mar}{vector of 4 positive numbers (graphical parameter)}

\item{xlab_1}{character (graphical parameter)}

\item{xlab_2}{character (graphical parameter)}

\item{ylab}{character (graphical parameter)}
}
\value{
makes a plot but does not return anything
}
\description{
Plot cell enrichment
}
