% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clisi.R
\name{plot_clisi}
\alias{plot_clisi}
\title{Making the plot for local enrichment}
\usage{
plot_clisi(
  local_1,
  local_2,
  col_vec = (scales::hue_pal())(nrow(local_1$common_clisi$membership_info)),
  l_bg = 75,
  c_bg = 50,
  alpha_bg = 0.5,
  xlab1 = "Distinct enrichment",
  xlab2 = "Distinct enrichment",
  ylab = "Common enrichment",
  main1 = "Modality 1",
  main2 = "Modality 2",
  ...
)
}
\arguments{
\item{local_1}{output of \code{clisi_information} on one Modality}

\item{local_2}{output of \code{clisi_information} on the other Modality}

\item{col_vec}{vector of colors}

\item{l_bg}{\code{l} parameter (luminosity, i.e., brightness) for individual cells}

\item{c_bg}{\code{c} parameter (chroma, i.e., color intensity) for individual cells}

\item{alpha_bg}{\code{alpha} parameter (color) for individual cells}

\item{xlab1}{\code{xlab} for Modality 1}

\item{xlab2}{\code{xlab} for Modality 2}

\item{ylab}{\code{ylab} for the shared modality}

\item{main1}{Title for plot corresponding to Modality 1}

\item{main2}{Title for plot corresponding to Modality 2}

\item{...}{extra parameters for \code{ggrepel::geom_text_repel}}
}
\value{
List of two \code{gg} objects
}
\description{
Making the plot for local enrichment
}
