% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clisi.R
\name{plot_clisi_legend}
\alias{plot_clisi_legend}
\title{Plot the cLISI legend}
\usage{
plot_clisi_legend(
  clisi_obj,
  col_vec = (scales::hue_pal())(nrow(clisi_obj$common_clisi$membership_info)),
  percent_coverage = 1,
  pch = 16,
  cex_point = 1,
  cex_text = 1,
  text_nudge = 0,
  xlim = c(0, 1),
  ...
)
}
\arguments{
\item{clisi_obj}{output of \code{clisi_information}}

\item{col_vec}{vector of colors}

\item{percent_coverage}{numeric}

\item{pch}{\code{pch} parameter}

\item{cex_point}{\code{cex} parameter for points}

\item{cex_text}{\code{cex} parameter for the text}

\item{text_nudge}{x-axis offset for the text}

\item{xlim}{\code{xlim} parameter for the plot}

\item{...}{additional graphical parameters}
}
\value{
nothing
}
\description{
Plot the cLISI legend
}
