% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data}
\alias{plot_data}
\title{Side-by-side UMAPs of the data}
\usage{
plot_data(obj, membership_vec, observed = F)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}

\item{membership_vec}{integer vector}

\item{observed}{boolean. This should \code{TRUE} if \code{obj} is the output of
\code{generate_data}, and you want to plot the "observed" data.
Otherwise, this is \code{FALSE} by default, meaning that if \code{obj} is the output of
\code{generate_data}, you are plotting the "true" data (i.e., not affected by noise),
or if \code{obj} is the output of \code{dcca_decomposition}, you are plotting
the estimated denoised observed matrix.}
}
\value{
shows a plot but returns nothing
}
\description{
Side-by-side UMAPs of the data
}
