% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embeddings}
\alias{plot_embeddings}
\title{Side-by-side UMAPs of the common, distinct and everything matrix}
\usage{
plot_embeddings(
  obj,
  membership_vec = NA,
  data_1 = T,
  data_2 = F,
  col_vec = (scales::hue_pal())(length(levels(membership_vec))),
  pca = F,
  only_embedding = F,
  metric = "cosine",
  main_addition = "",
  verbose = F
)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}

\item{membership_vec}{factor vector}

\item{data_1}{boolean}

\item{data_2}{boolean}

\item{col_vec}{vector of colors}

\item{pca}{boolean. If \code{TRUE}, plot the PCA embedding with the leading 2 components. 
If \code{FALSE}, plot the UMAP embedding.}

\item{only_embedding}{boolean}

\item{metric}{character for the distance metric used in \code{Seurat::RunUMAP},
typically \code{"cosine"} or \code{"euclidean"}}

\item{main_addition}{additional string to append to main of each plot}

\item{verbose}{boolean}
}
\value{
depends on \code{only_embedding}. 
If \code{TRUE}, returns three matrices as a list.
If \code{FALSE}, shows a plot but returns nothing
}
\description{
Side-by-side UMAPs of the common, distinct and everything matrix
}
