% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding2.R
\name{plot_embeddings2}
\alias{plot_embeddings2}
\title{Plot embeddings via frNN}
\usage{
plot_embeddings2(
  c_g,
  d_g,
  e_g,
  membership_vec = NA,
  col_vec = (scales::hue_pal())(length(levels(membership_vec))),
  only_embedding = F,
  main_addition = "",
  verbose = T,
  ...
)
}
\arguments{
\item{c_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the common embedding, where the non-zero entries represent distances}

\item{d_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the distinct embedding, where the non-zero entries represent distances}

\item{e_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the everything embedding, where the non-zero entries represent distances}

\item{membership_vec}{factor vector}

\item{col_vec}{vector of colors}

\item{only_embedding}{boolean}

\item{main_addition}{additional string to append to main of each plot}

\item{verbose}{boolean}

\item{...}{additional parameters for \code{Seurat:::RunUMAP.Graph}}
}
\value{
depends on \code{only_embedding}. 
If \code{TRUE}, returns three matrices as a list.
If \code{FALSE}, shows a plot but returns nothing
}
\description{
Plot embeddings via frNN
}
