% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmat}
\alias{plot_heatmat}
\title{Heatmap of the data}
\usage{
plot_heatmat(
  dat,
  luminosity = T,
  asp = nrow(dat)/ncol(dat),
  reserve_zero = T,
  ...
)
}
\arguments{
\item{dat}{matrix}

\item{luminosity}{boolean}

\item{asp}{numeric}

\item{reserve_zero}{boolean}

\item{...}{additional graphical parameters}
}
\value{
shows a plot but returns nothing
}
\description{
If \code{reserve_zero = T}, then reserve zero for white.
If \code{reserve_zero = F}, then all the greens are negative and all the reds are positive
}
