% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_laplacian.R
\name{plot_laplacian_variables}
\alias{plot_laplacian_variables}
\title{Plot the results of compute_smooth_signal}
\usage{
plot_laplacian_variables(
  val_vec,
  name_vec,
  factor_vec,
  col_vec,
  xlab = "Order of variables",
  ylab,
  main,
  baseline_col = "orange",
  text_cex = 2,
  ...
)
}
\arguments{
\item{val_vec}{vector of values to plot}

\item{name_vec}{name of the entries of \code{val_vec} as another vector of strings
of the same length}

\item{factor_vec}{vector of factors (among values \code{"0"}, \code{"1"} and \code{"2"})
of the same length as \code{val_vec} where entries with value \code{"1"}
represent the entries to be labeled via \code{ggrepel::geom_text_repel},
 entries with value \code{"2"} represent other entries to be colored (if any),
 and entries with value \code{"0"} all remaining entries}

\item{col_vec}{vector of 2 or 3 colors, depending on the number of levels in
\code{factor_vec}}

\item{xlab}{string}

\item{ylab}{string}

\item{main}{string}

\item{baseline_col}{color for the horizontal line}

\item{text_cex}{numeric}

\item{...}{extra parameters for \code{ggrepel::geom_text_repel}}
}
\value{
\code{gg} object
}
\description{
Plot the results of compute_smooth_signal
}
