% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softImpute.R
\name{plot_prediction_against_observed}
\alias{plot_prediction_against_observed}
\title{Plotting diagnostic to determine goodness of fit}
\usage{
plot_prediction_against_observed(
  diagnostic_mat,
  width = 0.8,
  scalar = NA,
  max_points = 5e+05,
  tol = 0.95,
  xlim = NA,
  ylim = NA,
  transparency = 0.2,
  cex_text = 1,
  ...
)
}
\arguments{
\item{diagnostic_mat}{one of the output matarices for \code{softImpute_diagnostic}}

\item{width}{parameter, controlling quantile of prediction region. The default is \code{0.8}, meaning
that the prediction region is by default from the 10th to 90th quantile.}

\item{scalar}{2 standard deviations for a Gaussian}

\item{max_points}{maximum number of points to be shown in the scatterplot, purely for visualization purposes only}

\item{tol}{parameter between \code{0} and \code{1} for how strict (\code{1} being the strictest) to measure
if the principal angle falls within the prediction region}

\item{xlim}{plotting parameter}

\item{ylim}{plotting parameter}

\item{transparency}{plotting parameter}

\item{cex_text}{plotting parameter}

\item{...}{additional plotting parameters}
}
\value{
a plot but nothing is returned to console
}
\description{
Plotting diagnostic to determine goodness of fit
}
