% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot summary of D-CCA}
\usage{
plot_summary(
  obj,
  xlab = "Latent dimension",
  ylab1 = "CCA objective",
  ylab2 = "Distinct \% for Modality 2",
  main = "",
  pch = 16,
  cex = 1,
  lwd = 1,
  ...
)
}
\arguments{
\item{obj}{output of \code{dcca_decomposition}}

\item{xlab}{\code{xlab} graphical parameter}

\item{ylab1}{\code{ylab} graphical parameter of left axis}

\item{ylab2}{\code{ylab} graphical parameter of right axis}

\item{main}{\code{main} graphical parameter}

\item{pch}{\code{pch} graphical parameter}

\item{cex}{\code{cex} graphical parameter}

\item{lwd}{\code{lwd} graphical parameter}

\item{...}{additional graphical parameter}
}
\value{
shows a plot but returns nothing
}
\description{
Plot summary of D-CCA
}
