% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_cell_enrichment.R
\name{postprocess_cell_enrichment.default}
\alias{postprocess_cell_enrichment.default}
\title{Computing cell enrichments for matrix object}
\usage{
\method{postprocess_cell_enrichment}{default}(
  input_obj,
  membership_vec,
  num_neigh,
  bool_cosine = T,
  bool_intersect = T,
  max_subsample = min(1000, length(membership_vec)),
  min_deg = 1,
  verbose = 0,
  ...
)
}
\arguments{
\item{input_obj}{matrix of \code{n} cells and \code{p} variable}

\item{membership_vec}{factor vector of the same length as the number of cells in \code{multiSVD}, denoting the
cell-types for each cell}

\item{num_neigh}{number of nearest neighbors}

\item{bool_cosine}{boolean, on whether or not to use cosine distance when constructing the SNN}

\item{bool_intersect}{boolean, on whether or not to symmetrize (via the AND function) the SNN}

\item{max_subsample}{maximum of cells to sample for each cell-type. If there are more than
\code{max_subsample} cells in a cell-type (dictated by \code{membership_vec}), a random
subset of cells will be selected for the sake of this function}

\item{min_deg}{minimum degree of each cell in the SNN}

\item{verbose}{non-negative integer}

\item{...}{additional arguments}
}
\description{
Computing cell enrichments for matrix object
}
