% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_cell_enrichment.R
\name{postprocess_cell_enrichment.multiSVD}
\alias{postprocess_cell_enrichment.multiSVD}
\title{Computing cell enrichments for multiSVD object}
\usage{
\method{postprocess_cell_enrichment}{multiSVD}(
  input_obj,
  membership_vec,
  max_subsample = min(1000, length(membership_vec)),
  verbose = 0,
  ...
)
}
\arguments{
\item{input_obj}{\code{multiSVD} object, after using \code{tiltedCCA_decomposition()}}

\item{membership_vec}{factor vector of the same length as the number of cells in \code{multiSVD}, denoting the
cell-types for each cell}

\item{max_subsample}{maximum of cells to sample for each cell-type. If there are more than
\code{max_subsample} cells in a cell-type (dictated by \code{membership_vec}), a random
subset of cells will be selected for the sake of this function}

\item{verbose}{non-negative integer}

\item{...}{additional arguments}
}
\description{
Computing cell enrichments for multiSVD object
}
