% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_cell_enrichment.R
\name{postprocess_cell_enrichment}
\alias{postprocess_cell_enrichment}
\title{Computing cell enrichments (generic)}
\usage{
postprocess_cell_enrichment(input_obj, ...)
}
\arguments{
\item{input_obj}{\code{multiSVD} object, after using \code{tiltedCCA_decomposition()}}

\item{...}{Additional arguments}
}
\description{
Computing cell enrichments (generic)
}
