% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_depvalue.R
\name{postprocess_marker_variables}
\alias{postprocess_marker_variables}
\title{Find the marker genes the separate one cell type from all other cell types}
\usage{
postprocess_marker_variables(
  de_list,
  log_thres_name = "avg_log2FC",
  log_thres = 2,
  num_quantile_threshold = 0.25,
  p_val_name = "p_val",
  p_val_adj_name = "p_val_adj",
  p_val_thres = 1e-04
)
}
\arguments{
\item{de_list}{output of the \code{tiltedCCA::differential_expression} function}

\item{log_thres_name}{character of the column in each of \code{de_list$de_list} for the log fold change}

\item{log_thres}{positive value, for selecting variables with log fold change larger than \code{log_thres}}

\item{num_quantile_threshold}{value between 0 and 1}

\item{p_val_name}{character of the column in each of \code{de_list$de_list} for the p-value}

\item{p_val_adj_name}{character of the column in each of \code{de_list$de_list} for the multiple-testing adjusted p-value}

\item{p_val_thres}{small positive value between 0 and 1, for selecting variables with p-values smaller than \code{p_val_thres}}
}
\value{
a list of variable vectors, one vector for each element in \code{de_list$level_vec}
}
\description{
Based on the output of \code{tiltedCCA::differential_expression}, for a
particular cell type, 1) when comparing against every other cell type,
find all the variables that have an adjusted p-value
smaller than \code{p_val_thres} and also have a log fold change larger than
\code{log_thres_name}, and 2) if the variable was selected for more than
\code{num_quantile_threshold} percentage of the comparisons with other cell 
types, we deem it as a marker gene for this cell type.
}
