% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helpers.R
\name{rotate_seurat_embeddings}
\alias{rotate_seurat_embeddings}
\title{Rotate embedding in a Seurat object}
\usage{
rotate_seurat_embeddings(seurat_obj, source_embedding, target_embedding)
}
\arguments{
\item{seurat_obj}{object of class \code{Seurat}}

\item{source_embedding}{character vector, so that \code{seurat_obj[[source_embedding]]} 
is a object of class \code{DimReduc}}

\item{target_embedding}{character vector, so that \code{seurat_obj[[target_embedding]]} 
is a object of class \code{DimReduc}}
}
\value{
an updated \code{seurat_obj}
}
\description{
This method rotates the embedding in \code{target_embedding}
to look (visually) most similar to \code{source_embedding}
}
