% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softImpute.R
\name{softImpute_diagnostic}
\alias{softImpute_diagnostic}
\title{softImpute diagnostic}
\usage{
softImpute_diagnostic(
  mat,
  K,
  num_val = ceiling(min(c(4, dim(mat)/10))),
  lambda = NA
)
}
\arguments{
\item{mat}{data matrix}

\item{K}{desired rank}

\item{num_val}{number to omit per row or per column}

\item{lambda}{numeric}
}
\value{
list
}
\description{
softImpute diagnostic
}
