% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltedCCA.R
\name{tiltedCCA_decomposition}
\alias{tiltedCCA_decomposition}
\title{Tilted-CCA Decomposition}
\usage{
tiltedCCA_decomposition(tiltedCCA_res, rank_c = NA, verbose = T)
}
\arguments{
\item{tiltedCCA_res}{output from \code{dcca_factor}}

\item{rank_c}{desired rank of cross-correlation matrix between \code{mat_1} and \code{mat_2} when running \code{dcca_factor}}

\item{verbose}{boolean}
}
\value{
list of class \code{dcca_decomp}
}
\description{
Tilted-CCA Decomposition
}
