% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clisi_information.R
\name{clisi_information}
\alias{clisi_information}
\title{Compute cLISI information}
\usage{
clisi_information(
  common_mat,
  distinct_mat,
  membership_vec,
  rank_c,
  rank_d,
  nn,
  frnn_approx = 0,
  radius_quantile = 0.95,
  min_subsample_cell = 50,
  verbose = T
)
}
\arguments{
\item{common_mat}{output from \code{dcca_decomposition}}

\item{distinct_mat}{output from \code{dcca_decomposition}}

\item{membership_vec}{factor}

\item{rank_c}{rank of \code{common_mat}}

\item{rank_d}{rank of \code{distinct_mat}}

\item{nn}{integer of number of nearest neighbors to determine the appropriate radius
for the frNN graph}

\item{frnn_approx}{small non-negative number}

\item{radius_quantile}{value between 0 and 1}

\item{min_subsample_cell}{positive integer, used for determining number of cells to 
compute cLISI for}

\item{verbose}{boolean}
}
\value{
two lists
}
\description{
Compute cLISI information
}
