% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clisi_information.R
\name{clisi_information}
\alias{clisi_information}
\title{Compute cLISI information}
\usage{
clisi_information(
  c_g,
  d_g,
  e_g,
  membership_vec,
  max_subsample_clisi = min(1000, nrow(c_g)),
  verbose = T
)
}
\arguments{
\item{c_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the common embedding, where the non-zero entries represent distances}

\item{d_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the distinct embedding, where the non-zero entries represent distances}

\item{e_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the everything embedding, where the non-zero entries represent distances}

\item{membership_vec}{factor vector}

\item{max_subsample_clisi}{positive integer, used for determining number of cells to 
compute cLISI for}

\item{verbose}{boolean}
}
\value{
three lists
}
\description{
Compute cLISI information
}
