% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frnn.R
\name{combine_frnn}
\alias{combine_frnn}
\title{Combining two frNN graphs}
\usage{
combine_frnn(
  dcca_obj,
  g_1,
  g_2,
  num_neigh = 30,
  center = T,
  common_1 = T,
  common_2 = T,
  keep_nn = T,
  renormalize = F,
  sampling_type = "adaptive_gaussian",
  verbose = 0
)
}
\arguments{
\item{dcca_obj}{output of \code{dcca_decomposition}}

\item{g_1}{graph object of class \code{dgCMatrix} where the non-zero entries represent distances
between nearest cells}

\item{g_2}{graph object of class \code{dgCMatrix} where the non-zero entries represent distances
between nearest cells}

\item{center}{boolean}

\item{common_1}{boolean}

\item{common_2}{boolean}

\item{keep_nn}{boolean}

\item{renormalize}{boolean}

\item{sampling_type}{string}

\item{verbose}{numeric from 0 to 2}

\item{nn}{integer of number of nearest neighbors to determine the appropriate radius
for the frNN graph}
}
\value{
object of class \code{dgCMatrix}
}
\description{
Combining two frNN graphs
}
