% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frnn.R
\name{construct_frnn}
\alias{construct_frnn}
\title{Construct fixed-radius NN graph}
\usage{
construct_frnn(
  obj,
  data_1 = T,
  data_2 = F,
  num_neigh = 30,
  bool_matrix = T,
  center = T,
  frnn_approx = 0,
  max_subsample_frnn = nrow(obj$common_score),
  membership_vec = NA,
  normalization_type = "none",
  radius_quantile = 0.5,
  renormalize = F,
  symmetrize = F,
  verbose = T
)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}

\item{data_1}{boolean}

\item{data_2}{boolean}

\item{bool_matrix}{boolean. If \code{TRUE}, output the graphs as a sparse matrix.
If \code{FALSE}, output the graphs as a list where each element of the list
corresponds with the element's neighbors}

\item{center}{boolean}

\item{frnn_approx}{small non-negative number}

\item{max_subsample_frnn}{positive integer, used for determining number of cells to 
compute cLISI for}

\item{membership_vec}{factor vector}

\item{normalization_type}{character among \code{"cosine_itself"},
\code{"cosine_everything"}, \code{"signac_itself"}, \code{"signac_everything"},
and \code{"none"}}

\item{radius_quantile}{value between 0 and 1}

\item{renormalize}{boolean}

\item{symmetrize}{boolean}

\item{verbose}{boolean}

\item{nn}{integer of number of nearest neighbors to determine the appropriate radius
for the frNN graph}
}
\value{
list, depends on \code{bool_matrix}
}
\description{
Construct fixed-radius NN graph
}
