% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frnn.R
\name{construct_frnn}
\alias{construct_frnn}
\title{Construct fixed-radius NN graph}
\usage{
construct_frnn(
  obj,
  nn,
  membership_vec,
  data_1 = T,
  data_2 = F,
  max_subsample_frnn = nrow(obj$common_score),
  frnn_approx = 0,
  radius_quantile = 0.5,
  bool_matrix = T,
  include_diag = T,
  verbose = T
)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}

\item{nn}{integer of number of nearest neighbors to determine the appropriate radius
for the frNN graph}

\item{membership_vec}{factor vector}

\item{data_1}{boolean}

\item{data_2}{boolean}

\item{max_subsample_frnn}{positive integer, used for determining number of cells to 
compute cLISI for}

\item{frnn_approx}{small non-negative number}

\item{radius_quantile}{value between 0 and 1}

\item{bool_matrix}{boolean. If \code{TRUE}, output the graphs as a sparse matrix.
If \code{FALSE}, output the graphs as a list where each element of the list
corresponds with the element's neighbors}

\item{include_diag}{boolean on whether or not the diagonal is included in the 
graph (currently only impactful if \code{bool_matrix=TRUE}).}

\item{verbose}{boolean}
}
\value{
list, depends on \code{bool_matrix}
}
\description{
Construct fixed-radius NN graph
}
