% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{dcca_decomposition}
\alias{dcca_decomposition}
\title{D-CCA Decomposition}
\usage{
dcca_decomposition(dcca_res, rank_c = NA, verbose = T)
}
\arguments{
\item{dcca_res}{output from \code{dcca_factor}}

\item{rank_c}{desired rank of cross-correlation matrix between \code{mat_1} and \code{mat_2} when running \code{dcca_factor}}

\item{verbose}{boolean}
}
\value{
list of class \code{dcca_decomp}
}
\description{
D-CCA Decomposition
}
