% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{dcca_factor}
\alias{dcca_factor}
\title{D-CCA Factorization}
\usage{
dcca_factor(
  mat_1,
  mat_2,
  rank_1,
  rank_2,
  meta_clustering = NA,
  num_neigh = max(round(nrow(mat_1)/20), 40),
  apply_shrinkage = T,
  reorthogonalize = F,
  verbose = T
)
}
\arguments{
\item{mat_1}{data matrix 1}

\item{mat_2}{data matrix 2}

\item{rank_1}{desired rank of data matrix 1}

\item{rank_2}{desired rank of data matrix 2}

\item{meta_clustering}{optional clustering}

\item{num_neigh}{number of neighbors to consider to computed the common percentage}

\item{apply_shrinkage}{boolean}

\item{reorthogonalize}{boolean}

\item{verbose}{boolean}
}
\value{
list of class \code{dcca}
}
\description{
D-CCA Factorization
}
