% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{dcca_factor}
\alias{dcca_factor}
\title{D-CCA Factorization}
\usage{
dcca_factor(
  mat_1,
  mat_2,
  dims_1,
  dims_2,
  center_1 = T,
  center_2 = T,
  meta_clustering = NA,
  num_neigh = max(round(nrow(mat_1)/20), 40),
  cell_max = nrow(mat_1),
  apply_shrinkage = T,
  fix_distinct_perc = F,
  verbose = T
)
}
\arguments{
\item{mat_1}{data matrix 1}

\item{mat_2}{data matrix 2}

\item{dims_1}{desired latent dimensions of data matrix 1}

\item{dims_2}{desired latent dimensions of data matrix 2}

\item{center_1}{boolean, on whether or not the center \code{mat_1} prior to SVD}

\item{center_2}{boolean, on whether or not the center \code{mat_2} prior to SVD}

\item{meta_clustering}{optional clustering}

\item{num_neigh}{number of neighbors to consider to computed the common percentage}

\item{cell_max}{number of cells used to compute the distinct percentaage}

\item{apply_shrinkage}{boolean}

\item{fix_distinct_perc}{boolean. If \code{TRUE}, the output \code{distinct_perc_2} will be fixed to at 0.5,
meaning the common scores will be the "middle" of \code{score_1} and \code{score_2}.
If \code{FALSE}, \code{distinct_perc_2} will be adaptively estimated via the
\code{.common_decomposition} function.}

\item{verbose}{boolean}
}
\value{
list of class \code{dcca}
}
\description{
D-CCA Factorization
}
