% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_decomposition.R
\name{.binary_search_radian}
\alias{.binary_search_radian}
\title{Find the appropriate radian for the common vector}
\usage{
.binary_search_radian(
  circle,
  left_radian,
  right_radian,
  distinct_perc_2,
  max_iter = 10,
  tol = 1e-06
)
}
\arguments{
\item{circle}{object from \code{.construct_circle}}

\item{left_radian}{radian (between \code{-pi} and \code{pi}) of the left-most vector}

\item{right_radian}{radian (between \code{-pi} and \code{pi}) of the right-most vector}

\item{distinct_perc_2}{the output of \code{.latent_distinct_perc_2}}

\item{max_iter}{positive integer}

\item{tol}{numeric}
}
\value{
radian (value between \code{right_radian} and \code{upper_radian})
}
\description{
Usage note: By \code{.representation_2d} works, the vector for
Modality 1 is always encoded by \code{right_radian}, and the
vector for Modality 2 is always encoded by \code{left_radian}.
}
\details{
If \code{distinct_perc_2} is close to 1, this means all the distinct information
(relatively speaking) lies in Modality 2. This means the returned
value should be very close to \code{right_radian}, meaning the distinct
vector for Modality 1 is near-0 (since the common vector is approximately
equal to the vector of Modality 1).
}
