% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{.cca}
\alias{.cca}
\title{Perform CCA}
\usage{
.cca(input_1, input_2, dims_1, dims_2, return_scores, tol = 1e-06)
}
\arguments{
\item{input_1}{first input}

\item{input_2}{second input}

\item{dims_1}{desired latent dimensions of data matrix 1. 
Only used if \code{input_1} is a matrix, not if it's  a list representing the SVD}

\item{dims_2}{desired latent dimensions of data matrix 2.
Only used if \code{input_2} is a matrix, not if it's  a list representing the SVD}

\item{return_scores}{boolean. If \code{TRUE}, return the scores (i.e., matrices where the rows are the cells).
If \code{FALSE}, return the loadings (i.e., matrices where the rows are the variables).
Either way, one of the output matrices will have \code{rank_1} columns and another
will have \code{rank_2} columns}

\item{tol}{small numeric}
}
\value{
list
}
\description{
This function takes either two matrices or two SVDs. Both \code{input_1}
and \code{input_2} must be "of the same type." Calls the \code{.compute_cca_aggregate_matrix} function.
}
