% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{.extract_umap_embedding}
\alias{.extract_umap_embedding}
\title{Extract UMAP embedding}
\usage{
.extract_umap_embedding(embedding, only_embedding, reduction_key = "UMAP")
}
\arguments{
\item{embedding}{return object from \code{.prepare_embeddings}}

\item{only_embedding}{boolean}

\item{reduction_key}{string for \code{Seurat::RunUMAP}}
}
\value{
list of three 2-column matrix or \code{Seurat} object, one for the
common, distinct and everything matrix
}
\description{
Extract UMAP embedding
}
