% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{.prepare_embeddings}
\alias{.prepare_embeddings}
\title{Prepare the appropriate inputs}
\usage{
.prepare_embeddings(obj, data_1, data_2, add_noise)
}
\arguments{
\item{obj}{output from either \code{generate_data} or \code{dcca_decomposition}}

\item{data_1}{boolean, for computing the embedding for data 1}

\item{data_2}{boolean, for computing the embedding for data 2}

\item{add_noise}{boolean, intended (if \code{TRUE}) to put the common and 
distinct scores "on the same scale" by adding appropriately-scaled Gaussian noise}
}
\value{
a list of 3 elements, representing the low-dimensional
representation for the common, distinct and everything matrices
for the intended dataset(s)
}
\description{
Prepare the appropriate inputs
}
