% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_explained_variance.R
\name{explained_variance}
\alias{explained_variance}
\title{Weights for cells and variables}
\usage{
explained_variance(
  dcca_decomp2,
  weight_func = function(x) {     x },
  verbose = T
)
}
\arguments{
\item{dcca_decomp2}{Output of \code{dcca_variance_decomposition}}

\item{weight_func}{function for scalars}

\item{verbose}{boolean}
}
\value{
4 vectors as a list
}
\description{
Weights for cells and variables
}
