% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{extract_embedding}
\alias{extract_embedding}
\title{Extract UMAP embedding}
\usage{
extract_embedding(
  obj,
  common_1 = T,
  common_2 = T,
  distinct_1 = T,
  distinct_2 = T,
  mode = "dcca",
  only_embedding = T,
  reduction_key = "UMAP"
)
}
\arguments{
\item{obj}{return object from \code{dcca_decomposition}}

\item{common_1}{boolean}

\item{common_2}{boolean}

\item{distinct_1}{boolean}

\item{distinct_2}{boolean}

\item{mode}{\code{"dcca"} or \code{"dmca"}}

\item{only_embedding}{boolean}

\item{reduction_key}{string for \code{Seurat::RunUMAP}}
}
\value{
2-column matrix
}
\description{
Extract UMAP embedding
}
