% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca.R
\name{extract_embedding}
\alias{extract_embedding}
\title{Extract UMAP embedding}
\usage{
extract_embedding(
  obj,
  common = T,
  distinct_1 = T,
  distinct_2 = T,
  only_embedding = T
)
}
\arguments{
\item{obj}{return object from \code{dcca_decomposition}}

\item{common}{boolean}

\item{distinct_1}{boolean}

\item{distinct_2}{boolean}

\item{only_embedding}{boolean}
}
\value{
2-column matrix
}
\description{
Extract UMAP embedding
}
