% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  svd_u_1,
  svd_u_2,
  svd_d_1,
  svd_d_2,
  svd_v_1,
  svd_v_2,
  num_neigh = max(round(nrow(svd_u_1)/20), 40),
  noise_val = 1
)
}
\arguments{
\item{svd_u_1}{\code{n} by \code{rank_1} orthogonal matrix}

\item{svd_u_2}{\code{n} by \code{rank_2} orthogonal matrix}

\item{svd_d_1}{vector of length \code{rank_1}}

\item{svd_d_2}{vector of length \code{rank_2}}

\item{svd_v_1}{\code{p1} by \code{rank_1} orthogonal matrix}

\item{svd_v_2}{\code{p2} by \code{rank_2} orthogonal matrix}

\item{num_neigh}{number of neighbors to consider to computed the common percentage}

\item{noise_val}{numeric for scalar of centered Gaussian noise}
}
\value{
list of outputs
}
\description{
Generate data
}
