% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  svd_u_1,
  svd_u_2,
  svd_d_1,
  svd_d_2,
  svd_v_1,
  svd_v_2,
  num_neigh = max(round(nrow(svd_u_1)/20), 40),
  noise_val = 1
)
}
\arguments{
\item{num_neigh}{number of neighbors to consider to computed the common percentage}

\item{noise_val}{numeric for scalar of centered Gaussian noise}

\item{score_1}{\code{n} by \code{rank_1} orthogonal matrix}

\item{score_2}{\code{n} by \code{rank_2} orthogonal matrix}

\item{coef_mat_1}{\code{rank_1} by \code{p_1} matrix}

\item{coef_mat_2}{\code{rank_2} by \code{p_2} matrix}
}
\value{
list of two matrices, one of dimension \code{n} by \code{p_1} and another of dimension \code{n} by \code{p_2}
}
\description{
Generate data
}
