% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clisi.R
\name{plot_clisi}
\alias{plot_clisi}
\title{Making the plot for cLISI}
\usage{
plot_clisi(
  clisi_1,
  clisi_2,
  col_vec = (scales::hue_pal())(nrow(clisi_1$common_clisi$membership_info)),
  cell_max = 5000,
  par_mar = c(4, 2.5, 0.5, 0.5),
  par_oma = c(0, 0, 2, 0),
  asp = T,
  pch_main = 16,
  cex_main = 1.5,
  pch_bg = 16,
  cex_bg = 1,
  alpha_bg = 0.5,
  l_bg = 95,
  c_bg = 50,
  xlim = c(0, 1),
  ylim = c(0, 1),
  gridsize = 5,
  col_grid = grDevices::rgb(0.8, 0.8, 0.8),
  lty_grid = 3,
  lwd_grid = 1,
  col_diag = "firebrick",
  lty_diag = 2,
  lwd_diag = 2,
  xlab1 = "Distinct information 1",
  xlab2 = "Distinct information 2",
  ylab = "Common information",
  ylab_dist = 0.5,
  main = "cLISI Information",
  cex_text_main = 1.5
)
}
\arguments{
\item{clisi_1}{output of \code{clisi_information} on one Modality}

\item{clisi_2}{output of \code{clisi_information} on the other Modality}

\item{col_vec}{vector of colors}

\item{cell_max}{number of cells of each cell-type to plot}

\item{par_mar}{\code{mar} parameter for \code{par}}

\item{par_oma}{\code{oma} parameter for \code{par}}

\item{asp}{graphical \code{asp} parameter}

\item{pch_main}{\code{pch} parameter for cell-types}

\item{cex_main}{\code{cex} parameter for cell-types}

\item{pch_bg}{\code{pch} parameter for individual cells}

\item{cex_bg}{\code{cex} parameter for individual cells}

\item{alpha_bg}{\code{alpha} parameter (color) for individual cells}

\item{l_bg}{\code{l} parameter (luminosity) for individual cells}

\item{c_bg}{\code{c} parameter (chroma, i.e., brightness) for individual cells}

\item{xlim}{\code{xlim} graphical parameter}

\item{ylim}{\code{ylim} graphical parameter}

\item{gridsize}{number of grid ticks}

\item{col_grid}{\code{col} for grid}

\item{lty_grid}{\code{lty} for grid}

\item{lwd_grid}{\code{lwd} for grid}

\item{col_diag}{\code{col} for diagonal line}

\item{lty_diag}{\code{lty} for diagonal line}

\item{lwd_diag}{\code{lwd} for diagonal line}

\item{xlab1}{\code{xlab} for Modality 1}

\item{xlab2}{\code{xlab} for Modality 2}

\item{ylab}{\code{ylab} for the shared modality}

\item{ylab_dist}{distance of \code{ylab} from y-axis}

\item{main}{\code{main} parameter}

\item{cex_text_main}{\code{cex} for \code{main} parameter}
}
\value{
nothing
}
\description{
See https://www.r-bloggers.com/2012/06/two-tips-adding-title-for-graph-with-multiple-plots-add-significance-asterix-onto-a-boxplot/
}
