% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embeddings}
\alias{plot_embeddings}
\title{Side-by-side UMAPs of the common, distinct and everything matrix}
\usage{
plot_embeddings(
  obj,
  membership_vec,
  data_1 = T,
  data_2 = F,
  add_noise = T,
  col_vec = (scales::hue_pal())(length(levels(membership_vec))),
  pca = F,
  only_embedding = F,
  main_addition = "",
  verbose = F
)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}

\item{membership_vec}{factor vector}

\item{data_1}{boolean}

\item{data_2}{boolean}

\item{add_noise}{boolean, intended (if \code{TRUE}) to put the common and 
distinct "on the same scale" by adding appropriately-scaled Gaussian noise}

\item{col_vec}{vector of colors}

\item{pca}{boolean. If \code{TRUE}, plot the PCA embedding with the leading 2 components. 
If \code{FALSE}, plot the UMAP embedding.}

\item{only_embedding}{boolean}

\item{main_addition}{additional string to append to main of each plot}

\item{verbose}{boolean}
}
\value{
shows a plot but returns nothing
}
\description{
Side-by-side UMAPs of the common, distinct and everything matrix
}
