% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_embeddings}
\alias{plot_embeddings}
\title{Side-by-side UMAPs of the common, distinct and everything matrix}
\usage{
plot_embeddings(obj, membership_vec, data_1 = T, data_2 = T)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}

\item{membership_vec}{integer vector}

\item{data_1}{boolean}

\item{data_2}{boolean}
}
\value{
shows a plot but returns nothing
}
\description{
Side-by-side UMAPs of the common, distinct and everything matrix
}
