% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding2.R
\name{plot_embeddings2}
\alias{plot_embeddings2}
\title{Plot embeddings via frNN}
\usage{
plot_embeddings2(
  dcca_obj,
  nn = 30,
  data_1 = T,
  data_2 = F,
  c_g = NA,
  d_g = NA,
  membership_vec = NA,
  col_vec = (scales::hue_pal())(length(levels(membership_vec))),
  only_embedding = F,
  main_addition = "",
  sampling_type = "adaptive_gaussian",
  keep_nn = T,
  verbose = T,
  ...
)
}
\arguments{
\item{dcca_obj}{output of \code{dcca_decomposition}}

\item{nn}{integer of number of nearest neighbors to determine the appropriate radius
for the frNN graph}

\item{data_1}{boolean}

\item{data_2}{boolean}

\item{c_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the common embedding, where the non-zero entries represent distances}

\item{d_g}{sparse matrix of class \code{dgCMatrix} from \code{construct_frnn}
representing the distinct embedding, where the non-zero entries represent distances}

\item{membership_vec}{factor vector}

\item{col_vec}{vector of colors}

\item{only_embedding}{boolean}

\item{main_addition}{additional string to append to main of each plot}

\item{sampling_type}{string}

\item{keep_nn}{boolean}

\item{verbose}{boolean}

\item{...}{additional parameters for \code{Seurat:::RunUMAP.Graph}}
}
\value{
depends on \code{only_embedding}. 
If \code{TRUE}, returns three matrices as a list.
If \code{FALSE}, shows a plot but returns nothing
}
\description{
Plot embeddings via frNN
}
