% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores}
\alias{plot_scores}
\title{Side-by-side plot of the canonical scores, colored by membership}
\usage{
plot_scores(obj, membership_vec)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}

\item{membership_vec}{integer vector}
}
\value{
shows a plot but returns nothing
}
\description{
Side-by-side plot of the canonical scores, colored by membership
}
