% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_heatmap}
\alias{plot_scores_heatmap}
\title{Side-by-side plot of the canonical scores as heatmaps}
\usage{
plot_scores_heatmap(obj)
}
\arguments{
\item{obj}{output of either \code{generate_data} or \code{dcca_decomposition}}
}
\value{
shows a plot but returns nothing
}
\description{
Side-by-side plot of the canonical scores as heatmaps
}
