% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltedCCA.R
\name{tiltedCCA}
\alias{tiltedCCA}
\title{Tilted-CCA Factorization}
\usage{
tiltedCCA(
  input_obj,
  discretization_gridsize = 21,
  enforce_boundary = F,
  fix_tilt_perc = F,
  verbose = 0
)
}
\arguments{
\item{input_obj}{\code{multiSVD} class, after creation via \code{compute_snns()}}

\item{discretization_gridsize}{positive integer for how many values between 0 and 1 (inclusive) to search the 
appropriate amount of tilt over}

\item{enforce_boundary}{boolean, on whether or not the tilt is required to stay between
the two canonical score vectors}

\item{fix_tilt_perc}{boolean or a numeric. If \code{FALSE}, then the tilt is adaptively
determined, and if \code{TRUE}, then the tilt is set to be equal to 
\code{0.5}. If numeric, the value should be between \code{0} and \code{1},
which the tilt will be set to.}

\item{verbose}{non-negative integer

For the tilt values (possibly set in \code{fix_tilt_perc}),
values close to 0 or 1 means the common space resembles the 
canonical scores of \code{mat_2} or \code{mat_1} respectively.}
}
\value{
updated \code{multiSVD} object
}
\description{
Computing the common-distinct decomposition via CCA, by setting
each latent dimension to the same common tilt.
}
