% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltedCCA.R
\name{tiltedCCA_decomposition}
\alias{tiltedCCA_decomposition}
\title{Tilted-CCA Decomposition}
\usage{
tiltedCCA_decomposition(input_obj, verbose = 0)
}
\arguments{
\item{verbose}{boolean}

\item{tiltedCCA_res}{output from \code{dcca_factor}}
}
\value{
list of class \code{dcca_decomp}
}
\description{
Tilted-CCA Decomposition
}
