% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltedCCA.R
\name{tiltedCCA_decomposition}
\alias{tiltedCCA_decomposition}
\title{Tilted-CCA Decomposition}
\usage{
tiltedCCA_decomposition(
  input_obj,
  bool_modality_1_full = T,
  bool_modality_2_full = T,
  verbose = 0
)
}
\arguments{
\item{input_obj}{\code{multiSVD} class, after creation via \code{tiltedCCA()} or \code{fine_tuning()}}

\item{bool_modality_1_full}{boolean, to compute the full-matrix decomposition for Modality 1 if \code{T}
or only the low-dimensional representation if \code{F}}

\item{bool_modality_2_full}{boolean, to compute the full-matrix decomposition for Modality 2 if \code{T}
or only the low-dimensional representation if \code{F}}

\item{verbose}{non-negative integer}
}
\value{
updated \code{multiSVD} object
}
\description{
After computing the appropriate tilts, recover the full decomposition of both modalities.
}
