% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltedCCA.R
\name{tiltedCCA_decomposition}
\alias{tiltedCCA_decomposition}
\title{Tilted-CCA Decomposition}
\usage{
tiltedCCA_decomposition(
  input_obj,
  bool_modality_1_full = T,
  bool_modality_2_full = T,
  verbose = 0
)
}
\arguments{
\item{input_obj}{output from \code{tiltedCCA}}

\item{bool_modality_1_full}{boolean}

\item{bool_modality_2_full}{boolean}

\item{verbose}{non-negative integer}
}
\value{
object of class \code{multiSVD}
}
\description{
Tilted-CCA Decomposition
}
